fix(migracion): chr() no acepta bigint, castear rn a int
row_number() devuelve bigint; 64 + rn quedaba bigint y chr(bigint) no
existe en Postgres (solo chr(integer)). Causaba que la migracion 0028
fallara siempre en produccion (transaccion se revertia entera, sin dejar
nada a medio aplicar -- confirmado por logs, funcion no existe: 42883).

+++ drizzle/0028_exotic_ironclad.sql
@@ -37,7 +37,7 @@ SELECT setval('product_code_seq', (SELECT COUNT(*) FROM "products"), true);--> s
 -- producto). Las variantes creadas DESPUES de esta migracion usan
 -- nextVariantCode() en la app, que si soporta mas de 26 (AA, AB...).
 UPDATE "product_variants" v
-SET "code" = p.code || '-' || chr(64 + sub.rn)
+SET "code" = p.code || '-' || chr((64 + sub.rn)::int)
 FROM (
   SELECT id, product_id, row_number() OVER (PARTITION BY product_id ORDER BY id) AS rn
   FROM "product_variants"
